Checkpoint: Complete rounding overhaul: all product quantities (annualUnits, monthly/quarterly delivery) rounded to integers with Math.floor. Milk remainder displayed as 'технологические потери' instead of warning. Products with annualUnits/12 < 1 grouped quarterly (delivered in March, June, September, December). Existing confirmed plans migrated with regenerated delivery schedules.

+++ server/routers/productTrack.ts
@@ -101,7 +101,7 @@ const deleteProductOptionInput = z.object({
 /* Tier-based selection input */
 const tierSelectionInput = z.object({
   catalogItemId: z.number().int().positive(),
-  annualUnits: z.number().min(0).max(100_000),
+  annualUnits: z.number().int().min(0).max(100_000),
 });
 
 const deliveryStatusInput = z.object({
